+++ armadillo/src/main/java/at/favre/lib/armadillo/Armadillo.java
@@ -2,11 +2,15 @@ package at.favre.lib.armadillo;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.os.Build;
 import android.provider.Settings;
 import android.support.annotation.Nullable;
 
 import java.security.Provider;
 import java.security.SecureRandom;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 
 import at.favre.lib.bytes.Bytes;
@@ -21,6 +25,8 @@ import at.favre.lib.bytes.Bytes;
 public final class Armadillo {
 
     public static final int CONTENT_KEY_OUT_BYTE_LENGTH = 20;
+    public static final int DEFAULT_PROTOCOL_VERSION = 0;
+    public static final int KITKAT_PROTOCOL_VERSION = -19;
 
     private Armadillo() {
     }
@@ -68,19 +74,15 @@ public final class Armadillo {
 
         private EncryptionFingerprint fingerprint;
         private StringMessageDigest stringMessageDigest = new HkdfMessageDigest(BuildConfig.PREF_SALT, CONTENT_KEY_OUT_BYTE_LENGTH);
-        @AuthenticatedEncryption.KeyStrength
-        private int keyStrength = AuthenticatedEncryption.STRENGTH_HIGH;
-        private AuthenticatedEncryption authenticatedEncryption;
-        private KeyStretchingFunction keyStretchingFunction = new ArmadilloBcryptKeyStretcher();
-        private DataObfuscator.Factory dataObfuscatorFactory = new HkdfXorObfuscator.Factory();
+        private EncryptionProtocolConfig.Builder defaultConfig = EncryptionProtocolConfig.newDefaultConfig();
+        private List<EncryptionProtocolConfig> additionalDecryptionConfigs = new ArrayList<>(2);
         private SecureRandom secureRandom = new SecureRandom();
         private RecoveryPolicy recoveryPolicy = new RecoveryPolicy.Default(true, false);
         private char[] password;
         private boolean supportVerifyPassword = false;
         private Provider provider;
-        private int cryptoProtocolVersion = 0;
         private boolean enableDerivedPasswordCache = false;
-        private Compressor compressor = new DisabledCompressor();
+        private boolean enableKitKatSupport = false;
 
         private Builder(SharedPreferences sharedPreferences) {
             this(sharedPreferences, null, null);
@@ -99,7 +101,7 @@ public final class Armadillo {
         /**
          * The encryption fingerprint is in important security measure. When no user password is
          * provided, it is the most important source of entropy to derive the key for the encryption.
-         *
+         * <p>
          * Set the default fingerprint using sources explained in {@link EncryptionFingerprintFactory}.
          *
          * @param context used to gather sources from the Android OS
@@ -112,14 +114,14 @@ public final class Armadillo {
         /**
          * The encryption fingerprint is in important security measure. When no user password is
          * provided, it is the most important source of entropy to derive the key for the encryption.
-         *
+         * <p>
          * Set the default fingerprint using sources explained in {@link EncryptionFingerprintFactory
          * with addtional custom data. Setting this is <strong>highly recommended</strong> as it makes
          * it more difficult for an attacker calculate the key the more random the input is.
-         *
+         * <p>
          * See the README.md for explainating on what to use as additionalData.
          *
-         * @param context used to gather sources from the Android OS
+         * @param context        used to gather sources from the Android OS
          * @param additionalData provided additional custom data
          * @return builder
          */
@@ -130,17 +132,17 @@ public final class Armadillo {
         /**
          * The encryption fingerprint is in important security measure. When no user password is
          * provided, it is the most important source of entropy to derive the key for the encryption.
-         *
+         * <p>
          * Set the default fingerprint using sources explained in {@link EncryptionFingerprintFactory
          * with addtional custom data. Setting this is <strong>highly recommended</strong> as it makes
          * it more difficult for an attacker calculate the key the more random the input is.
-         *
+         * <p>
          * This is the same as {@link #encryptionFingerprint(Context, byte[])} but accepts strings
          * instead of a byte array.
-         *
+         * <p>
          * See the README.md for explainating on what to use as additionalData.
          *
-         * @param context used to gather sources from the Android OS
+         * @param context        used to gather sources from the Android OS
          * @param additionalData provided additional custom data
          * @return builder
          */
@@ -161,7 +163,7 @@ public final class Armadillo {
         /**
          * The encryption fingerprint is in important security measure. When no user password is
          * provided, it is the most important source of entropy to derive the key for the encryption.
-         *
+         * <p>
          * Provide a fully custom fingerprint implementation (or instance). Use this if you don't
          * agree with the default implementation.
          * <hr />
@@ -179,7 +181,7 @@ public final class Armadillo {
         /**
          * The encryption fingerprint is in important security measure. When no user password is
          * provided, it is the most important source of entropy to derive the key for the encryption.
-         *
+         * <p>
          * Provide a fully custom fingerprint byte array. Use this if you don't
          * agree with the default implementation.
          * <hr />
@@ -198,16 +200,16 @@ public final class Armadillo {
          * The content key digest is responsible for hashing the key in the key-value pair of
          * a shared preference. E.g. if the key is "name" and the value "Bob", the key "name" will
          * be hashed before it is persisted to disk.
-         *
+         * <p>
          * This method will alter the salt used for that hash. Setting this is highly recommended, since
          * it will change the default hashes of the key (so that somebody else's "name" key, won't
          * hash to the exact same output). Recommended value: use the AndroidID, as it will be different
          * on every app install from SDK 26+. See {@link Settings.Secure#ANDROID_ID}.
-         *
+         * <p>
          * Note that changing the salt will make old data inaccessible, since the key won't match
          * anymore.
          *
-         * @param salt to be used for content key hash
+         * @param salt to be used for content key hash (should be > 16 byte)
          * @return builder
          */
         public Builder contentKeyDigest(byte[] salt) {
@@ -218,7 +220,7 @@ public final class Armadillo {
          * The content key digest is responsible for hashing the key in the key-value pair of
          * a shared preference. E.g. if the key is "name" and the value "Bob", the key "name" will
          * be hashed before it is persisted to disk.
-         *
+         * <p>
          * This is a more advanced setting. Per default a hash will be {@link #CONTENT_KEY_OUT_BYTE_LENGTH}
          * bytes long. If you think that is too long (wasting space) or too small (not enough entropy)
          * modify the length with this config.
@@ -236,7 +238,7 @@ public final class Armadillo {
          * The content key digest is responsible for hashing the key in the key-value pair of
          * a shared preference. E.g. if the key is "name" and the value "Bob", the key "name" will
          * be hashed before it is persisted to disk.
-         *
+         * <p>
          * Use this to set a fully custom implementation of the digest.
          * <hr />
          * <strong>Note:</strong> <em>Only set if you know what you are doing.</em>
@@ -245,8 +247,7 @@ public final class Armadillo {
          * @return builder
          */
         public Builder contentKeyDigest(StringMessageDigest stringMessageDigest) {
-            Objects.requireNonNull(stringMessageDigest);
-            this.stringMessageDigest = stringMessageDigest;
+            this.stringMessageDigest = Objects.requireNonNull(stringMessageDigest);
             return this;
         }
 
@@ -263,13 +264,13 @@ public final class Armadillo {
          * <em>Note:</em> Usually there is no real advantage to set it to VERY HIGH as HIGH (128 bit key
          * length) is fully secure for the foreseeable future. VERY HIGH only adds more security margin
          * for possible quantum computer attacks (but if you are a user which is threatened by these
-         * kinds of attacks you wouldn't use this lib anyway).
+         * kinds of attacks you probably require higher degrees af protection).
          *
          * @param keyStrength HIGH (default) or VERY HIGH
          * @return builder
          */
         public Builder encryptionKeyStrength(@AuthenticatedEncryption.KeyStrength int keyStrength) {
-            this.keyStrength = keyStrength;
+            defaultConfig.keyStrength(keyStrength);
             return this;
         }
 
@@ -299,8 +300,7 @@ public final class Armadillo {
          * @return builder
          */
         public Builder symmetricEncryption(AuthenticatedEncryption authenticatedEncryption) {
-            Objects.requireNonNull(authenticatedEncryption);
-            this.authenticatedEncryption = authenticatedEncryption;
+            defaultConfig.authenticatedEncryption(Objects.requireNonNull(authenticatedEncryption));
             return this;
         }
 
@@ -315,8 +315,7 @@ public final class Armadillo {
          * @return builder
          */
         public Builder keyStretchingFunction(KeyStretchingFunction keyStretchingFunction) {
-            Objects.requireNonNull(keyStretchingFunction);
-            this.keyStretchingFunction = keyStretchingFunction;
+            defaultConfig.keyStretchingFunction(Objects.requireNonNull(keyStretchingFunction));
             return this;
         }
 
@@ -330,8 +329,7 @@ public final class Armadillo {
          * @return builder
          */
         public Builder dataObfuscatorFactory(DataObfuscator.Factory dataObfuscatorFactory) {
-            Objects.requireNonNull(dataObfuscatorFactory);
-            this.dataObfuscatorFactory = dataObfuscatorFactory;
+            defaultConfig.dataObfuscatorFactory(Objects.requireNonNull(dataObfuscatorFactory));
             return this;
         }
 
@@ -413,7 +411,7 @@ public final class Armadillo {
          * @return builder
          */
         public Builder cryptoProtocolVersion(int version) {
-            this.cryptoProtocolVersion = version;
+            defaultConfig.protocolVersion(version);
             return this;
         }
 
@@ -459,10 +457,62 @@ public final class Armadillo {
         /**
          * Compresses the content with given compressor before encrypting and writing it to shared preference.
          *
+         * @param compressor to set
          * @return builder
          */
         public Builder compress(Compressor compressor) {
-            this.compressor = compressor;
+            defaultConfig.compressor(compressor);
+            return this;
+        }
+
+        /**
+         * Add new {@link EncryptionProtocolConfig} to be added to the supported decryption-config
+         * list. That means, if you have encrypted data with an older encryption config you may add
+         * it here to be able to <strong>decrypt</strong> it again. These configs are however never
+         * used for <strong>encryption</strong>.
+         * <p>
+         * To match configs {@link EncryptionProtocolConfig#protocolVersion} is used, therefore if
+         * you add a config here it must match the persisted protocolVersion (Armadillo default is
+         * {@link #DEFAULT_PROTOCOL_VERSION})
+         * <p>
+         * This may be used for migration of encryption protocols.
+         *
+         * @param config to be added to the list
+         * @return builder
+         */
+        public Builder addAdditionalDecryptionProtocolConfig(EncryptionProtocolConfig config) {
+            additionalDecryptionConfigs.add(config);
+            return this;
+        }
+
+        /**
+         * Clear additionalDecryptionConfigs list.
+         * See also {@link #addAdditionalDecryptionProtocolConfig(EncryptionProtocolConfig)}.
+         *
+         * @return builder
+         */
+        public Builder clearAdditionalDecryptionProtocolConfigs() {
+            additionalDecryptionConfigs.clear();
+            return this;
+        }
+
+        /**
+         * Manually enable kitkatSupport. Unfortunately Android SDK 19 (KITKAT) does not fully
+         * support AES GCM mode. Therefore a backwards compatible implementation of AES
+         * (see {@link AesCbcEncryption}) which uses CBC + Encrypt-then-mac which should have
+         * a similar security strength.
+         * <p>
+         * The backwards compatible implementation will <strong>only</strong> be used if
+         * {@link Build.VERSION#SDK_INT} is lower or equal to 19, the default version otherwise.
+         * Additionally this implementation will be added to the additionalDecryptionConfigs so you
+         * can still decrypt the old content after upgrading to lollipop and above (uses protocol-version
+         * {@link Armadillo#KITKAT_PROTOCOL_VERSION} to mark.
+         *
+         * @param enable kitkat support
+         * @return builder
+         */
+        public Builder enableKitKatSupport(boolean enable) {
+            enableKitKatSupport = enable;
             return this;
         }
 
@@ -476,17 +526,49 @@ public final class Armadillo {
                 throw new IllegalArgumentException("No encryption fingerprint is set - see encryptionFingerprint() methods");
             }
 
-            if (authenticatedEncryption == null) {
-                authenticatedEncryption = new AesGcmEncryption(secureRandom, provider);
+            EncryptionProtocolConfig config = defaultConfig.build();
+
+            if (enableKitKatSupport) {
+                if (config.authenticatedEncryption != null) {
+                    throw new IllegalStateException("enabling kitkat support will prevent using custom encryption implementation");
+                }
+
+                @SuppressWarnings("deprecation")
+                EncryptionProtocolConfig kitkatSupportConfig = EncryptionProtocolConfig.newBuilder(config)
+                        .authenticatedEncryption(new AesCbcEncryption(secureRandom, provider))
+                        .protocolVersion(KITKAT_PROTOCOL_VERSION)
+                        .build();
+
+                additionalDecryptionConfigs.add(kitkatSupportConfig);
+
+                // set current encryption config to kitkat support if on kitkat device
+                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+                    config = kitkatSupportConfig;
+                }
             }
 
-            EncryptionProtocol.Factory factory = new DefaultEncryptionProtocol.Factory(cryptoProtocolVersion, fingerprint, stringMessageDigest, authenticatedEncryption, keyStrength,
-                keyStretchingFunction, dataObfuscatorFactory, secureRandom, enableDerivedPasswordCache, compressor);
+            if (config.authenticatedEncryption == null) {
+                config = EncryptionProtocolConfig.newBuilder(config)
+                        .authenticatedEncryption(new AesGcmEncryption(secureRandom, provider)).build();
+            }
+
+            EncryptionProtocol.Factory factory = new DefaultEncryptionProtocol.Factory(config,
+                    fingerprint, stringMessageDigest, secureRandom, enableDerivedPasswordCache, Collections.unmodifiableList(additionalDecryptionConfigs));
+
+            checkKitKatSupport(config.authenticatedEncryption);
 
             if (sharedPreferences != null) {
                 return new SecureSharedPreferences(sharedPreferences, factory, recoveryPolicy, password, supportVerifyPassword);
             } else {
                 return new SecureSharedPreferences(context, prefName, factory, recoveryPolicy, password, supportVerifyPassword);
+            }
+        }
+
+        private void checkKitKatSupport(AuthenticatedEncryption authenticatedEncryption) {
+            if (Build.VERSION.SDK_INT == Build.VERSION_CODES.KITKAT &&
+                    authenticatedEncryption.getClass().equals(AesGcmEncryption.class)) {
+                throw new UnsupportedOperationException("aes gcm is not supported with KitKat, add support " +
+                        "manually with Armadillo.Builder.enableKitKatSupport()");
             }
         }
     }